# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "contact": {}
    },
    "paths": {
...
        "/namespaces": {
            "get": {
                "description": "Retrieves a list of namespaces from the Kubernetes cluster.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "namespaces"
                ],
                "summary": "List namespaces",
                "parameters": [
                    {
                        "type": "string",
                        "default": "Bearer \u003cAdd access token here\u003e",
                        "description": "Insert your access token",
                        "name": "Authorization",
                        "in": "header",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Maximum number of namespaces to retrieve",
                        "name": "limit",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Pagination token for fetching more namespaces",
                        "name": "continue",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "List of namespaces",
                        "schema": {
                            "$ref": "#/definitions/controller.SuccessResponse"
                        }
                    },
                    "500": {
                        "description": "Bad request or error message",
                        "schema": {
                            "$ref": "#/definitions/controller.FailureResponse"
                        }
                    }
                }
            },
            "put": {
                "description": "Updates an existing namespace in the Kubernetes cluster.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "namespaces"
                ],
                "summary": "Update an existing namespace",
                "parameters": [
                    {
                        "type": "string",
                        "default": "Bearer \u003cAdd access token here\u003e",
                        "description": "Insert your access token",
                        "name": "Authorization",
                        "in": "header",
                        "required": true
                    },
                    {
                        "description": "Namespace request body",
                        "name": "namespace",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/model.NamespaceUpdateRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successfully updated the namespace",
                        "schema": {
                            "$ref": "#/definitions/controller.SuccessResponse"
                        }
                    },
                    "400": {
                        "description": "Bad request or invalid data",
                        "schema": {
                            "$ref": "#/definitions/controller.FailureResponse"
                        }
                    },
                    "500": {
                        "description": "Interval error",
                        "schema": {
                            "$ref": "#/definitions/controller.FailureResponse"
                        }
                    }
                }
            },
            "post": {
                "description": "Creates a new namespace in the Kubernetes cluster.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "namespaces"
                ],
                "summary": "Create a new namespace",
                "parameters": [
                    {
                        "type": "string",
                        "default": "Bearer \u003cAdd access token here\u003e",
                        "description": "Insert your access token",
                        "name": "Authorization",
                        "in": "header",
                        "required": true
                    },
                    {
                        "description": "Namespace request body",
                        "name": "namespace",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/model.NamespaceCreateRequest"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Successfully created namespace",
                        "schema": {
                            "$ref": "#/definitions/controller.SuccessResponse"
                        }
                    },
                    "400": {
                        "description": "Bad request or error message",
                        "schema": {
                            "$ref": "#/definitions/controller.FailureResponse"
                        }
                    },
                    "500": {
                        "description": "Interval error",
                        "schema": {
                            "$ref": "#/definitions/controller.FailureResponse"
                        }
                    }
                }
            }
        },
        "/namespaces/{id}": {
            "get": {
                "description": "Retrieves a namespace from the Kubernetes cluster by its name or UID.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "namespaces"
                ],
                "summary": "Get a namespace by name or UID",
                "parameters": [
                    {
                        "type": "string",
                        "default": "Bearer \u003cAdd access token here\u003e",
                        "description": "Insert your access token",
                        "name": "Authorization",
                        "in": "header",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Name or UID of the namespace",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Details of the requested namespace",
                        "schema": {
                            "$ref": "#/definitions/controller.SuccessResponse"
                        }
                    },
                    "500": {
                        "description": "Interval error",
                        "schema": {
                            "$ref": "#/definitions/controller.FailureResponse"
                        }
                    }
                }
            },
            "delete": {
                "description": "Deletes a namespace from the Kubernetes cluster by its name or UID.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "namespaces"
                ],
                "summary": "Delete a namespace by name or UID",
                "parameters": [
                    {
                        "type": "string",
                        "default": "Bearer \u003cAdd access token here\u003e",
                        "description": "Insert your access token",
                        "name": "Authorization",
                        "in": "header",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Name or UID of the namespace",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success message",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "500": {
                        "description": "Interval error",
                        "schema": {
                            "$ref": "#/definitions/controller.FailureResponse"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "controller.FailureResponse": {
            "type": "object",
            "properties": {
                "error": {
                    "type": "string"
                },
                "message": {
                    "type": "string"
                }
            }
        },
        "controller.SuccessResponse": {
            "type": "object",
            "properties": {
                "data": {},
                "message": {
                    "type": "string"
                }
            }
        },
...
        "model.CreateOptions": {
            "type": "object",
            "properties": {
                "apiVersion": {
                    "type": "string"
                },
                "dryRun": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "fieldManager": {
                    "type": "string"
                },
                "fieldValidation": {
                    "type": "string"
                },
                "kind": {
                    "type": "string"
                }
            }
        },
...
        "model.FinalizerName": {
            "type": "string",
            "enum": [
                "kubernetes"
            ],
            "x-enum-varnames": [
                "FinalizerKubernetes"
            ]
        },
...
        "model.Namespace": {
            "type": "object",
            "properties": {
                "apiVersion": {
                    "type": "string"
                },
                "kind": {
                    "type": "string"
                },
                "metadata": {
                    "$ref": "#/definitions/model.ObjectMeta"
                },
                "spec": {
                    "description": "Spec defines the behavior of the Namespace.\n+optional",
                    "allOf": [
                        {
                            "$ref": "#/definitions/model.NamespaceSpec"
                        }
                    ]
                },
                "status": {
                    "description": "Status describes the current status of a Namespace.\n+optional",
                    "allOf": [
                        {
                            "$ref": "#/definitions/model.NamespaceStatus"
                        }
                    ]
                }
            }
        },
        "model.NamespaceCondition": {
            "type": "object",
            "properties": {
                "lastTransitionTime": {
                    "description": "+optional",
                    "type": "string"
                },
                "message": {
                    "description": "+optional",
                    "type": "string"
                },
                "reason": {
                    "description": "+optional",
                    "type": "string"
                },
                "status": {
                    "description": "Status of the condition, one of True, False, Unknown.",
                    "type": "string"
                },
                "type": {
                    "description": "Type of namespace controller condition.",
                    "allOf": [
                        {
                            "$ref": "#/definitions/model.NamespaceConditionType"
                        }
                    ]
                }
            }
        },
        "model.NamespaceConditionType": {
            "type": "string",
            "enum": [
                "NamespaceDeletionDiscoveryFailure",
                "NamespaceDeletionContentFailure",
                "NamespaceDeletionGroupVersionParsingFailure",
                "NamespaceContentRemaining",
                "NamespaceFinalizersRemaining"
            ],
            "x-enum-varnames": [
                "NamespaceDeletionDiscoveryFailure",
                "NamespaceDeletionContentFailure",
                "NamespaceDeletionGVParsingFailure",
                "NamespaceContentRemaining",
                "NamespaceFinalizersRemaining"
            ]
        },
        "model.NamespaceCreateRequest": {
            "type": "object",
            "properties": {
                "namespace": {
                    "$ref": "#/definitions/model.Namespace"
                },
                "opts": {
                    "$ref": "#/definitions/model.CreateOptions"
                }
            }
        },
        "model.NamespacePhase": {
            "type": "string",
            "enum": [
                "Active",
                "Terminating"
            ],
            "x-enum-varnames": [
                "NamespaceActive",
                "NamespaceTerminating"
            ]
        },
        "model.NamespaceSpec": {
            "type": "object",
            "properties": {
                "finalizers": {
                    "description": "Finalizers is an opaque list of values that must be empty to permanently remove object from storage.\n+optional",
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/model.FinalizerName"
                    }
                }
            }
        },
        "model.NamespaceStatus": {
            "type": "object",
            "properties": {
                "conditions": {
                    "description": "Represents the latest available observations of a namespace's current state.",
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/model.NamespaceCondition"
                    }
                },
                "phase": {
                    "description": "Phase is the current lifecycle phase of the namespace.\n+optional",
                    "allOf": [
                        {
                            "$ref": "#/definitions/model.NamespacePhase"
                        }
                    ]
                }
            }
        },
        "model.NamespaceUpdateRequest": {
            "type": "object",
            "properties": {
                "namespace": {
                    "$ref": "#/definitions/model.Namespace"
                },
                "opts": {
                    "$ref": "#/definitions/model.UpdateOptions"
                }
            }
        },
        "model.ObjectMeta": {
            "type": "object",
            "properties": {
                "annotations": {
                    "type": "object",
                    "additionalProperties": {
                        "type": "string"
                    }
                },
                "creationTimestamp": {
                    "type": "string"
                },
                "deletionGracePeriodSeconds": {
                    "type": "integer"
                },
                "deletionTimestamp": {
                    "type": "string"
                },
                "finalizers": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "generateName": {
                    "type": "string"
                },
                "generation": {
                    "type": "integer"
                },
                "labels": {
                    "type": "object",
                    "additionalProperties": {
                        "type": "string"
                    }
                },
                "name": {
                    "type": "string"
                },
                "namespace": {
                    "type": "string"
                },
                "ownerReferences": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/model.OwnerReference"
                    }
                },
                "resourceVersion": {
                    "type": "string"
                },
                "uid": {
                    "type": "string"
                }
            }
        },
        "model.OwnerReference": {
            "type": "object",
            "properties": {
                "apiVersion": {
                    "type": "string"
                },
                "blockOwnerDeletion": {
                    "type": "boolean"
                },
                "controller": {
                    "type": "boolean"
                },
                "kind": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                }
            }
        },
...
        "model.UpdateOptions": {
            "type": "object",
            "properties": {
                "apiVersion": {
                    "type": "string"
                },
                "dryRun": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "fieldManager": {
                    "type": "string"
                },
                "fieldValidation": {
                    "type": "string"
                },
                "kind": {
                    "type": "string"
                }
            }
        },
...
    }
}